now resizes properly
