+++ src/info_handling.py
@@ -2,7 +2,7 @@ import os, sys, yt, typing
 from typing import Any
 from abc import ABC, abstractmethod
 from enum import Enum
-from v_3_options import *
+from options import *
 
 class EventBroker:
     """
